+++ codec/cbor/codec_test.go
@@ -1,0 +1,86 @@
+package cbor_test
+
+import (
+	"reflect"
+
+	. "github.com/dogmatiq/dogma/fixtures"
+	. "github.com/dogmatiq/marshalkit/codec/cbor"
+	. "github.com/dogmatiq/marshalkit/codec/internal/fixtures"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("type Codec", func() {
+	var codec *Codec
+
+	BeforeEach(func() {
+		codec = &Codec{}
+	})
+
+	Describe("func Query()", func() {
+		It("uses the unqualified type-name as the portable type", func() {
+			rt := reflect.TypeOf(MessageA{})
+
+			caps := codec.Query(
+				[]reflect.Type{rt},
+			)
+
+			Expect(caps.Types[rt]).To(Equal("MessageA"))
+		})
+
+		It("uses the user-defined type name", func() {
+			type LocalMessage MessageA
+			rt := reflect.TypeOf(LocalMessage{})
+
+			caps := codec.Query(
+				[]reflect.Type{rt},
+			)
+
+			Expect(caps.Types[rt]).To(Equal("LocalMessage"))
+		})
+
+		It("uses the element name for pointer types", func() {
+			var m **MessageA
+			rt := reflect.TypeOf(m)
+
+			caps := codec.Query(
+				[]reflect.Type{rt},
+			)
+
+			Expect(caps.Types[rt]).To(Equal("MessageA"))
+		})
+	})
+
+	Describe("func BasicMediaType()", func() {
+		It("returns the expected basic media-type", func() {
+			Expect(codec.BasicMediaType()).To(Equal("application/cbor"))
+		})
+	})
+
+	Describe("func Marshal()", func() {
+		It("marshals the value", func() {
+			data, err := codec.Marshal(
+				&ProtoMessage{
+					Value: "<value>",
+				},
+			)
+			Expect(err).ShouldNot(HaveOccurred())
+			Expect(data).To(Equal([]byte("\xa1evalueg<value>")))
+		})
+	})
+
+	Describe("func Unmarshal()", func() {
+		It("unmarshals the data", func() {
+			data := []byte("\xa1evalueg<value>")
+
+			m := &ProtoMessage{}
+			err := codec.Unmarshal(data, m)
+			Expect(err).ShouldNot(HaveOccurred())
+			Expect(m).To(Equal(
+				&ProtoMessage{
+					Value: "<value>",
+				},
+			))
+		})
+	})
+})