+++ fixtures/marshaler.go
@@ -6,6 +6,7 @@ import (
 	"github.com/dogmatiq/dogma/fixtures"
 	"github.com/dogmatiq/marshalkit"
 	"github.com/dogmatiq/marshalkit/codec"
+	"github.com/dogmatiq/marshalkit/codec/cbor"
 	"github.com/dogmatiq/marshalkit/codec/json"
 )
 
@@ -47,6 +48,7 @@ func init() {
 		},
 		[]codec.Codec{
 			&json.Codec{},
+			&cbor.Codec{},
 		},
 	)
 	if err != nil {